JPA Repositories added

+++ src/main/java/com/PatManSystem/main/Repository/PatientRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PatientRepository extends JpaRepository<Patientinformation, Long> {
+public interface PatientRepository extends JpaRepository<Patientinformation, Integer> {
 }
